requirements.txt
openai-whisper
pydub
numpy
jiwer
pytest
black
boto3
python-dotenv
